Added support for SOCKS proxies.
This adds a new proxy_type config file option that has the following values:

- http - HTTP proxy, the default. Equivalent to the previous behavior.
- socks4 - SOCKSv4
- socks4a - SOCKSv4a
- socks5 - SOCKSv5

For socks4 and socks4a the proxy username is required to be specified as well,
either interactively on the console or in the secrets file.

In addition to the config file, the support for SOCKS proxies has been added
to the (http|https|all)_proxy environment variables, where the proxy type
is determined by the URI scheme. The scheme is mandatory (and it has been
before, despite the documentation).

Updated documentation and config file accordingly.